Move team_abbrev to odds_core.team to fix scheduler Lambda import crash
storage/writers.py imported team_abbrev from oddsportal_common, which has
a top-level `from oddsharvester.core.scrape_result import ScrapeResult`.
The scheduler Lambda (Dockerfile.lambda) doesn't bundle oddsharvester,
so any import chain reaching oddsportal_common crashes the handler at
module load time with "No module named 'oddsharvester'".

Introduced by #281 which added the team_abbrev import to writers.py.
